+++ core/pageable-core/src/main/java/com/github/aqiu202/page/PageableInput.java
@@ -11,7 +11,7 @@ public interface PageableInput {
 
     int getPageSize();
 
-    long getOffset();
+    int getOffset();
 
     boolean hasPrevious();
